Support SCRAM-SHA-1-PLUS.  Add GSASL_CB_TLS_UNIQUE.  Use GnuTLS for tls-unique CB.

+++ lib/scram/printer.c
@@ -131,8 +131,8 @@ scram_print_server_first (struct scram_server_first *sf, char **out)
   if (!scram_valid_server_first (sf))
     return -1;
 
-  n = asprintf (out, "r=%s,s=%s,i=%d",
-		sf->nonce, sf->salt, sf->iter);
+  n = asprintf (out, "r=%s,s=%s,i=%lu",
+		sf->nonce, sf->salt, (unsigned long) sf->iter);
   if (n <= 0 || *out == NULL)
     return -1;
